refactor(command): move types into multiple sub-folders
