+++ src/features/discord/messages/functions/commands/checks/discord-contains-this-command-with-one-of-these-prefixes.spec.ts
@@ -1,6 +1,6 @@
 import { createMock } from "ts-auto-mock";
 import { DiscordMessageCommandEnum } from "../../../enums/command/discord-message-command.enum";
-import { IDiscordContainsThisCommandWithOneOfThesePrefixesData } from "../../../interfaces/commands/discord-contains-this-command-with-one-of-these-prefixes-data";
+import { IDiscordContainsThisCommandWithOneOfThesePrefixesData } from "../../../interfaces/commands/checks/discord-contains-this-command-with-one-of-these-prefixes-data";
 import { discordContainsThisCommandWithOneOfThesePrefixes } from "./discord-contains-this-command-with-one-of-these-prefixes";
 
 describe(`containsThisCommandWithOneOfThesePrefixes()`, (): void => {
