+++ src/features/discord/messages/functions/commands/checks/discord-strictly-contains-this-command-with-prefix.ts
@@ -1,4 +1,4 @@
-import { IDiscordStrictlyContainsThisCommandWithPrefixData } from "../../../interfaces/commands/discord-strictly-contains-this-command-with-prefix-data";
+import { IDiscordStrictlyContainsThisCommandWithPrefixData } from "../../../interfaces/commands/checks/discord-strictly-contains-this-command-with-prefix-data";
 import { discordGetCommandRegexp } from "../regexp/discord-get-command-regexp";
 import xregexp from "xregexp";
