+++ src/features/discord/messages/functions/commands/getters/discord-get-command-and-first-argument.ts
@@ -1,6 +1,6 @@
 import _ from "lodash";
-import { IDiscordExtractFromCommandCallbackData } from "../../../interfaces/commands/discord-extract-from-command-callback-data";
-import { IDiscordGetCommandAndFirstArgumentData } from "../../../interfaces/commands/discord-get-command-and-first-argument-data";
+import { IDiscordExtractFromCommandCallbackData } from "../../../interfaces/commands/checks/discord-extract-from-command-callback-data";
+import { IDiscordGetCommandAndFirstArgumentData } from "../../../interfaces/commands/getters/discord-get-command-and-first-argument-data";
 import { discordExtractFromCommand } from "../checks/discord-extract-from-command";
 import { discordGetCommandWithFirstArgumentRegexp } from "../regexp/discord-get-command-with-first-argument-regexp";
 import { discordGetFormattedMessage } from "../formatters/discord-get-formatted-message";
