+++ src/features/discord/messages/functions/commands/getters/discord-get-command-first-argument.spec.ts
@@ -1,6 +1,6 @@
 import { createMock } from "ts-auto-mock";
 import { DiscordMessageCommandEnum } from "../../../enums/command/discord-message-command.enum";
-import { IDiscordGetCommandFirstArgumentData } from "../../../interfaces/commands/discord-get-command-first-argument-data";
+import { IDiscordGetCommandFirstArgumentData } from "../../../interfaces/commands/getters/discord-get-command-first-argument-data";
 import { discordGetCommandFirstArgument } from "./discord-get-command-first-argument";
 
 describe(`discordGetCommandFirstArgument()`, (): void => {
