+++ src/features/discord/messages/functions/commands/getters/discord-get-command-flags.ts
@@ -1,7 +1,7 @@
 import _ from "lodash";
 import xregexp, { ExecArray } from "xregexp";
-import { IDiscordExtractFromCommandCallbackData } from "../../../interfaces/commands/discord-extract-from-command-callback-data";
-import { IDiscordGetCommandFlagsData } from "../../../interfaces/commands/discord-get-command-flags-data";
+import { IDiscordExtractFromCommandCallbackData } from "../../../interfaces/commands/checks/discord-extract-from-command-callback-data";
+import { IDiscordGetCommandFlagsData } from "../../../interfaces/commands/getters/discord-get-command-flags-data";
 import { discordExtractFromCommand } from "../checks/discord-extract-from-command";
 import { discordGetCommandWithFirstArgumentAndFlagsRegexp } from "../regexp/discord-get-command-with-first-argument-and-flags-regexp";
 import { discordGetFormattedMessage } from "../formatters/discord-get-formatted-message";
