+++ src/features/discord/messages/functions/commands/getters/discord-get-command-with-prefix.ts
@@ -1,5 +1,5 @@
 import _ from "lodash";
-import { IDiscordGetCommandWithPrefix } from "../../../interfaces/commands/discord-get-command-with-prefix";
+import { IDiscordGetCommandWithPrefix } from "../../../interfaces/commands/getters/discord-get-command-with-prefix";
 
 export function discordGetCommandWithPrefix({
   prefix,
