+++ src/features/discord/messages/functions/commands/regexp/discord-get-command-with-first-argument-and-flags-regexp.spec.ts
@@ -1,7 +1,7 @@
 import { createMock } from "ts-auto-mock";
 import xregexp from "xregexp";
 import { DiscordMessageCommandEnum } from "../../../enums/command/discord-message-command.enum";
-import { IDiscordGetCommandWithFirstArgumentAndFlagsRegexpData } from "../../../interfaces/commands/discord-get-command-with-first-argument-and-flags-regexp-data";
+import { IDiscordGetCommandWithFirstArgumentAndFlagsRegexpData } from "../../../interfaces/commands/getters/discord-get-command-with-first-argument-and-flags-regexp-data";
 import { discordGetCommandWithFirstArgumentAndFlagsRegexp } from "./discord-get-command-with-first-argument-and-flags-regexp";
 
 describe(`discordGetCommandWithFirstArgumentAndFlagsRegexp()`, (): void => {
