+++ src/features/discord/messages/interfaces/commands/checks/discord-contains-this-command-with-prefix-data.ts
@@ -1,4 +1,4 @@
-import { DiscordMessageCommandEnum } from "../../enums/command/discord-message-command.enum";
+import { DiscordMessageCommandEnum } from "../../../enums/command/discord-message-command.enum";
 
 export interface IDiscordContainsThisCommandWithPrefixData {
   commands: DiscordMessageCommandEnum | DiscordMessageCommandEnum[];
