+++ src/features/discord/messages/interfaces/commands/checks/discord-extract-from-command-callback-data.ts
@@ -1,4 +1,4 @@
-import { DiscordMessageCommandEnum } from "../../enums/command/discord-message-command.enum";
+import { DiscordMessageCommandEnum } from "../../../enums/command/discord-message-command.enum";
 
 export interface IDiscordExtractFromCommandCallbackData {
   command: DiscordMessageCommandEnum;
