+++ src/features/discord/messages/interfaces/commands/checks/discord-extract-from-command-data.ts
@@ -1,4 +1,4 @@
-import { DiscordMessageCommandEnum } from "../../enums/command/discord-message-command.enum";
+import { DiscordMessageCommandEnum } from "../../../enums/command/discord-message-command.enum";
 import { IDiscordExtractFromCommandCallbackData } from "./discord-extract-from-command-callback-data";
 
 export interface IDiscordExtractFromCommandData {
