+++ src/features/discord/messages/interfaces/commands/checks/discord-has-this-command-data.ts
@@ -1,4 +1,4 @@
-import { DiscordMessageCommandEnum } from "../../enums/command/discord-message-command.enum";
+import { DiscordMessageCommandEnum } from "../../../enums/command/discord-message-command.enum";
 
 export interface IDiscordHasThisCommandData {
   commands: DiscordMessageCommandEnum | DiscordMessageCommandEnum[];
