+++ src/features/discord/messages/interfaces/commands/getters/discord-get-command-with-first-argument-and-flags-regexp-data.ts
@@ -1,4 +1,4 @@
-import { DiscordMessageCommandEnum } from "../../enums/command/discord-message-command.enum";
+import { DiscordMessageCommandEnum } from "../../../enums/command/discord-message-command.enum";
 
 export interface IDiscordGetCommandWithFirstArgumentAndFlagsRegexpData {
   command: DiscordMessageCommandEnum;
